customise swagger docs and cleanup
