--- src/main/java/com/assignment/customerdatamanagement/exceptions/CustomerNotFoundException.java
@@ -1,15 +1,0 @@
-package com.assignment.customerdatamanagement.exceptions;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-
-@Getter
-@AllArgsConstructor
-public class CustomerNotFoundException extends RuntimeException {
-
-    private final String httpStatusCode;
-    private final String message;
-
-
-}
