--- src/main/java/com/assignment/customerdatamanagement/models/CustomerAddress.java
@@ -1,31 +1,0 @@
-package com.assignment.customerdatamanagement.models;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Data
-public class CustomerAddress {
-
-    @NotBlank(message = "HouseNumber number cannot be empty")
-    private String houseNumber;
-
-    @NotBlank(message = "Street cannot be empty")
-    private String street;
-
-    @NotBlank(message = "City cannot be empty")
-    private String city;
-
-    @NotBlank(message = "Country is mandatory")
-    private String country;
-
-    @NotBlank(message = "Postal code is mandatory")
-    private String postalCode;
-}
-
