--- src/main/java/com/assignment/customerdatamanagement/security/SecurityConfig.java
@@ -1,60 +1,0 @@
-package com.assignment.customerdatamanagement.security;
-
-import lombok.SneakyThrows;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Value("${basicauth.userName}")
-    private String basicUserName;
-    @Value("${basicauth.password}")
-    private String basicPassword;
-
-    @SneakyThrows
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        return http
-                .cors(Customizer.withDefaults())
-                .csrf(AbstractHttpConfigurer::disable)
-                .authorizeHttpRequests(auth -> auth
-                        .requestMatchers(HttpMethod.POST, "/customers").authenticated()
-                        .requestMatchers(HttpMethod.PUT, "/customers/{customerId}").authenticated()
-                        .requestMatchers(HttpMethod.GET, "/customers/**").authenticated()
-                        .requestMatchers("/swagger-ui/**","/api-docs.html","/api-docs/**").permitAll()
-                        .anyRequest().denyAll())
-                .httpBasic(httpSecurityHttpBasicConfigurer -> httpSecurityHttpBasicConfigurer
-                        .realmName("basic"))
-                .build();
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService(BCryptPasswordEncoder bCryptPasswordEncoder) {
-        InMemoryUserDetailsManager manager = new InMemoryUserDetailsManager();
-        manager.createUser(
-                User.withUsername(basicUserName)
-                        .password(bCryptPasswordEncoder.encode(basicPassword))
-                        .build()
-        );
-        return manager;
-    }
-
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-}
-
