+++ src/main/java/com/crm/customer/customerdatamanagement/exceptions/CustomerNotFoundException.java
@@ -1,0 +1,15 @@
+package com.crm.customer.customerdatamanagement.exceptions;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+
+@Getter
+@AllArgsConstructor
+public class CustomerNotFoundException extends RuntimeException {
+
+    private final String httpStatusCode;
+    private final String message;
+
+
+}
