+++ src/main/java/com/crm/customer/customerdatamanagement/exceptions/ErrorResponse.java
@@ -1,0 +1,13 @@
+package com.crm.customer.customerdatamanagement.exceptions;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+import java.util.List;
+
+@AllArgsConstructor
+@Getter
+public class ErrorResponse {
+    private String errorCode;
+    private List<String> errorMessages;
+}
