+++ src/main/java/com/crm/customer/customerdatamanagement/models/CustomerAddress.java
@@ -1,0 +1,31 @@
+package com.crm.customer.customerdatamanagement.models;
+
+import jakarta.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+public class CustomerAddress {
+
+    @NotBlank(message = "HouseNumber number cannot be empty")
+    private String houseNumber;
+
+    @NotBlank(message = "Street cannot be empty")
+    private String street;
+
+    @NotBlank(message = "City cannot be empty")
+    private String city;
+
+    @NotBlank(message = "Country is mandatory")
+    private String country;
+
+    @NotBlank(message = "Postal code is mandatory")
+    private String postalCode;
+}
+
