+++ src/main/resources/application.yml
@@ -21,14 +21,14 @@ spring:
   threads:
     virtual:
       enabled: true
-      
+
 springdoc:
   api-docs:
     path: /api-docs
   swagger-ui:
-    path : /api-docs.html
+    path: /api-docs.html
     operationsSorter: method
-    
+
 # in production Credential  will  be read from secure place like credhub
 basicauth:
   userName: Non-Personal-Account
