--- src/test/java/com/assignment/customerdatamanagement/CustomerDataManagementApplicationTests.java
@@ -1,16 +1,0 @@
-package com.assignment.customerdatamanagement;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.ApplicationContext;
-
-@SpringBootTest
-class CustomerDataManagementApplicationTests {
-
-    @Test
-    void contextLoads(ApplicationContext applicationContext) {
-        Assertions.assertNotNull(applicationContext);
-    }
-
-}
