+++ src/test/java/com/crm/customer/customerdatamanagement/CustomerDataManagementApplicationTests.java
@@ -1,0 +1,16 @@
+package com.crm.customer.customerdatamanagement;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.ApplicationContext;
+
+@SpringBootTest
+class CustomerDataManagementApplicationTests {
+
+    @Test
+    void contextLoads(ApplicationContext applicationContext) {
+        Assertions.assertNotNull(applicationContext);
+    }
+
+}
